requirements.txt
Django==2.2.10
djangorestframework==3.12.1
drf-nested-routers==0.92.1
pytz==2020.1
sqlparse==0.4.1
